transferred legacy classes into a separate package, added league support for NewPlayer class, changed rational computations to work with longs instead of ints and removed obsolete copying operations from Graph getters.
