+++ src/GraphXings/Game/Match/NewMatch.java
@@ -31,7 +31,10 @@ public class NewMatch
 	 * The time limit for each game in the match.
 	 */
 	private long gameTimeLimit;
-
+	/**
+	 * If true, print all game results (true by default).
+	 */
+	private boolean isVerbose;
 	/**
 	 * Constructs a match object.
 	 * @param player1 The first player.
@@ -46,6 +49,7 @@ public class NewMatch
 		this.bestOutOf = bestOutOf;
 		this.factory = factory;
 		this.gameTimeLimit = Long.MAX_VALUE;
+		isVerbose = true;
 	}
 
 	/**
@@ -63,6 +67,7 @@ public class NewMatch
 		this.bestOutOf = bestOutOf;
 		this.factory = factory;
 		this.gameTimeLimit = gameTimeLimit;
+		isVerbose = true;
 	}
 
 	/**
@@ -81,7 +86,10 @@ public class NewMatch
 			NewGame game = new NewGame(gi.getG(),gi.getWidth(),gi.getHeight(),player1,player2,gameTimeLimit);
 			NewGameResult gr = game.play();
 			gamesPlayed++;
-			System.out.println("Game " + gamesPlayed + ": " + gr.announceResult());
+			if (isVerbose)
+			{
+				System.out.println("Game " + gamesPlayed + ": " + gr.announceResult());
+			}
 			if (gamesPlayed > 2*bestOutOf)
 			{
 				end = true;
@@ -110,5 +118,14 @@ public class NewMatch
 			}
 		}
 		return new NewMatchResult(player1,player2,gamesWon1,gamesWon2);
+	}
+
+	/**
+	 * Sets whether or not all intermediate game results should be displayed.
+	 * @param verbose True, if intermediate results should be displayed, false otherwise.
+	 */
+	public void setVerbose(boolean verbose)
+	{
+		this.isVerbose = verbose;
 	}
 }
