+++ src/GraphXings/Legacy/Algorithms/RandomPlayer.java
@@ -1,0 +1,97 @@
+package GraphXings.Legacy.Algorithms;
+
+import GraphXings.Data.Coordinate;
+import GraphXings.Data.Graph;
+import GraphXings.Data.Vertex;
+import GraphXings.Game.GameMove;
+
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Random;
+
+/**
+ * A player performing random moves.
+ */
+public class RandomPlayer implements Player
+{
+    /**
+     * The name of the random player.
+     */
+    private String name;
+    /**
+     * A random number generator.
+     */
+    private Random r;
+
+    /**
+     * Creates a random player with the assigned name.
+     * @param name
+     */
+    public RandomPlayer(String name)
+    {
+        this.name = name;
+        this.r =  new Random(name.hashCode());
+    }
+
+    @Override
+    public GameMove maximizeCrossings(Graph g, HashMap<Vertex, Coordinate> vertexCoordinates, List<GameMove> gameMoves, int[][] usedCoordinates, HashSet<Vertex> placedVertices, int width, int height)
+    {
+        return randomMove(g,usedCoordinates,placedVertices,width,height);
+    }
+
+    @Override
+    public GameMove minimizeCrossings(Graph g, HashMap<Vertex, Coordinate> vertexCoordinates, List<GameMove> gameMoves, int[][] usedCoordinates, HashSet<Vertex> placedVertices, int width, int height)
+    {
+        return randomMove(g,usedCoordinates,placedVertices,width,height);
+    }
+
+    @Override
+    public void initializeNextRound(Graph g, int width, int height, Role role)
+    {
+
+    }
+
+    /**
+     * Computes a random valid move.
+     * @param g The graph.
+     * @param usedCoordinates The used coordinates.
+     * @param placedVertices The already placed vertices.
+     * @param width The width of the game board.
+     * @param height The height of the game board.
+     * @return A random valid move.
+     */
+    private GameMove randomMove(Graph g, int[][] usedCoordinates, HashSet<Vertex> placedVertices, int width, int height)
+    {
+        int stillToBePlaced = g.getN()- placedVertices.size();
+        int next = r.nextInt(stillToBePlaced);
+        int skipped = 0;
+        Vertex v=null;
+        for (Vertex u : g.getVertices())
+        {
+            if (!placedVertices.contains(u))
+            {
+                if (skipped < next)
+                {
+                    skipped++;
+                    continue;
+                }
+                v=u;
+                break;
+            }
+        }
+        Coordinate c = new Coordinate(0,0);
+        do
+        {
+            c = new Coordinate(r.nextInt(width),r.nextInt(height));
+        }
+        while (usedCoordinates[c.getX()][c.getY()]!=0);
+        return new GameMove(v,c);
+    }
+
+    @Override
+    public String getName()
+    {
+        return name;
+    }
+}
